impl group member

+++ bot/line/line_utilities.py
@@ -6,7 +6,7 @@ from bot.line.line_settings import api as line_api
 from bot.models import Group, LineGroup, LineUser, User
 
 
-def get_user_by_line_user_id_from_database(line_user_id):
+def get_user_by_line_user_id_from_database(line_user_id: str)->User:
     '''LINEのユーザーIDでデータベースからユーザーを取得する。ない場合は作成する'''
     try:
         return User.objects.get(line_user__user_id__exact=line_user_id)
@@ -15,7 +15,7 @@ def get_user_by_line_user_id_from_database(line_user_id):
             "ユーザー(LineUserID: {})が見つかりませんでした。".format(line_user_id))
 
 
-def get_group_by_line_group_id_from_database(line_group_id):
+def get_group_by_line_group_id_from_database(line_group_id: str)->Group:
     '''LINEのグループIDでデータベースからグループを取得する'''
     try:
         return Group.objects.get(line_group__group_id__exact=line_group_id)
@@ -24,7 +24,7 @@ def get_group_by_line_group_id_from_database(line_group_id):
             "グループ(LineGroupID: {})が見つかりませんでした。".format(line_group_id))
 
 
-def register_user_by_line_user_id(line_user_id):
+def register_user_by_line_user_id(line_user_id: str)->User:
     '''LINEユーザーIDでユーザーを登録する。戻り値は新しいユーザーデータ'''
     user_profile = line_api.get_profile(line_user_id)
     name = user_profile.display_name
@@ -46,13 +46,20 @@ def register_user_by_line_user_id(line_user_id):
         new_line_user.delete()
         raise
 
-def register_user_by_line_user_in_group_id(line_user_id, line_group_id):
-    '''LINEユーザーIDでユーザーを登録する。戻り値は新しいユーザーデータ'''
-    user_profile = line_api.get_group_member_profile(line_group_id, line_user_id)
-    name = user_profile.display_name
-    # LINEユーザーをデータベースに登録
-    new_line_user = LineUser.objects.create(user_id=line_user_id, name=name)
+
+def register_user_by_line_user_id_in_group(line_user_id: str, line_group_id: str)->User:
+    '''LINEユーザーIDとLINEグループIDででユーザーを登録する。
+    LINEIDに紐付いたグループが作成されている必要があり、紐付いているグループにもユーザーが登録される。
+    戻り値は新しいユーザーデータ。'''
+    new_user = None
+    new_line_user = None
     try:
+        user_profile = line_api.get_group_member_profile(
+            line_group_id, line_user_id)
+        name = user_profile.display_name
+        # LINEユーザーをデータベースに登録
+        new_line_user = LineUser.objects.create(
+            user_id=line_user_id, name=name)
         # ユーザをデータベースに登録
         counter = 1
         name_candidate = name
@@ -62,14 +69,21 @@ def register_user_by_line_user_in_group_id(line_user_id, line_group_id):
             counter += 1
         new_user = User.objects.create(name=name_candidate, line_user=new_line_user,
                                        authority=UserAuthority.Watcher.name)
+        # グループにユーザーを登録
+        group = get_group_by_line_group_id_from_database(line_group_id)
+        group.members.add(new_user)
+        group.save()
         print("ユーザー(LineID: {}, Name: {})をデータベースに登録しました。".format(line_user_id, name))
         return new_user
     except Exception:
-        new_line_user.delete()
+        if new_line_user:
+            new_line_user.delete()
+        if new_line_user:
+            new_user.delete()
         raise
 
 
-def register_group_by_line_group_id(line_group_id):
+def register_group_by_line_group_id(line_group_id: str)->Group:
     '''LINEユーザーIDでユーザーを登録する。戻り値は新しいユーザーデータ。
     グループ名はグループ数から自動で「グループ**」と付けられる。'''
     # LINEグループをデータベースに登録
@@ -87,3 +101,11 @@ def register_group_by_line_group_id(line_group_id):
     except Exception:
         new_line_group.delete()
         raise
+
+
+def check_and_add_member(user: User, group: Group):
+    '''ユーザーがグループに属している確認して、属していないなら登録する。'''
+    if not group.members.filter(id=user.id).exists():
+        print("ユーザー「{}」をグループ「{}」に登録。".format(user.name, group.name))
+        group.members.add(user)
+        group.save()
